+++ test/unit/label.c
@@ -1,4 +1,6 @@
 #include "unit.h"
+#include <float.h>
+#include <math.h>
 
 #define uiLabelPtrFromState(s) uiControlPtrFromState(uiLabel, s)
 
@@ -66,6 +68,49 @@ static void labelSetTextEmptyString(void **state)
 	uiFreeText(rv);
 }
 
+static void labelFontSize(void **state)
+{
+	uiLabel **l = uiLabelPtrFromState(state);
+	double size;
+
+	*l = uiNewLabel("Text");
+	size = uiLabelFontSize(*l);
+	assert_true(size > 0 && size <= DBL_MAX);
+}
+
+static void labelSetFontSize(void **state)
+{
+	uiLabel **l = uiLabelPtrFromState(state);
+
+	*l = uiNewLabel("Text");
+	uiLabelSetFontSize(*l, 18.5);
+	assert_true(fabs(uiLabelFontSize(*l) - 18.5) < 0.000001);
+}
+
+static void labelSetTextPreservesFontSize(void **state)
+{
+	uiLabel **l = uiLabelPtrFromState(state);
+
+	*l = uiNewLabel("Text");
+	uiLabelSetFontSize(*l, 18.5);
+	uiLabelSetText(*l, "Changed");
+	assert_true(fabs(uiLabelFontSize(*l) - 18.5) < 0.000001);
+}
+
+static void labelResetFontSize(void **state)
+{
+	uiLabel **l = uiLabelPtrFromState(state);
+	double initial;
+
+	*l = uiNewLabel("Text");
+	initial = uiLabelFontSize(*l);
+	uiLabelSetFontSize(*l, initial + 5);
+	uiLabelResetFontSize(*l);
+	assert_true(fabs(uiLabelFontSize(*l) - initial) < 0.000001);
+	uiLabelResetFontSize(*l);
+	assert_true(fabs(uiLabelFontSize(*l) - initial) < 0.000001);
+}
+
 #define labelUnitTest(f) cmocka_unit_test_setup_teardown((f), \
 		unitTestSetup, unitTestTeardown)
 
@@ -78,8 +123,11 @@ int labelRunUnitTests(void)
 		labelUnitTest(labelTextEmptyString),
 		labelUnitTest(labelSetText),
 		labelUnitTest(labelSetTextEmptyString),
+		labelUnitTest(labelFontSize),
+		labelUnitTest(labelSetFontSize),
+		labelUnitTest(labelSetTextPreservesFontSize),
+		labelUnitTest(labelResetFontSize),
 	};
 
 	return cmocka_run_group_tests_name("uiLabel", tests, unitTestsSetup, unitTestsTeardown);
 }
-